PlayerAI: fix reset()
Mostly I was just fixing indentation, bracket placement and spacing to
make it consistent with the rest of the file.  Since we never defined
any sort of project-wide style rules (or at least, no one -told- me
about any), it'd be awesome if people respected my preferences, as I try
to do to theirs.  *sigh*

git-svn-id: svn+ssh://fileserverness.ath.cx/serv/svn/virus-td@261 05766cc9-4f33-4ba7-801d-bd015708efd9